+++ stubs/DecodeLabs/Prophet.php
@@ -30,11 +30,26 @@ class Prophet implements Proxy
     public static function getRepository(): Ref1 {
         return static::$instance->getRepository();
     }
+    public static function tryLoadAssistant(Ref2|string $blueprint, string $serviceName): ?Ref3 {
+        return static::$instance->tryLoadAssistant(...func_get_args());
+    }
     public static function loadAssistant(Ref2|string $blueprint, string $serviceName): Ref3 {
         return static::$instance->loadAssistant(...func_get_args());
     }
+    public static function loadFreshAssistant(Ref2|string $blueprint, string $serviceName): Ref3 {
+        return static::$instance->loadFreshAssistant(...func_get_args());
+    }
+    public static function deleteAssistant(Ref2|string $blueprint, string $serviceName): bool {
+        return static::$instance->deleteAssistant(...func_get_args());
+    }
+    public static function tryLoadThread(Ref2|string $blueprint, Ref4 $subject): ?Ref5 {
+        return static::$instance->tryLoadThread(...func_get_args());
+    }
     public static function loadThread(Ref2|string $blueprint, Ref4 $subject): Ref5 {
         return static::$instance->loadThread(...func_get_args());
+    }
+    public static function deleteThread(Ref2|string $blueprint, Ref4 $subject): bool {
+        return static::$instance->deleteThread(...func_get_args());
     }
     public static function serializeThreadWithMessages(Ref5 $thread, ?string $afterId = NULL, int $limit = 20): array {
         return static::$instance->serializeThreadWithMessages(...func_get_args());